requirements.txt
websockets
aiohttp
python-dotenv
numpy
